% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pass.R
\name{create_pass}
\alias{create_pass}
\title{create_pass}
\usage{
create_pass(local = TRUE)
}
\arguments{
\item{local}{TRUE or FALSE either local windows machine or workbench remote}
}
\value{
"pass created"
}
\description{
generate the authinfo file to connect SAS
}
\examples{
\dontrun{
create_pass(local = TRUE)
} 
}
