% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sascfg_personal_py.R
\name{create_sascfg_personal_py}
\alias{create_sascfg_personal_py}
\title{create_sascfg_personal_py}
\usage{
create_sascfg_personal_py(
  local = TRUE,
  iomhost = "aws-saspv1.mycompany.com",
  class_id = "letter-number"
)
}
\arguments{
\item{local}{TRUE or FALSE either local windows machine or workbench remote}

\item{iomhost}{proc iomoperate - url of Operating System Services}

\item{class_id}{proc iomoperate - SAS Workspace Server Class identifier information}
}
\value{
"sascfg_personal.py created"
}
\description{
generate the config files to connect SAS
}
\examples{
\dontrun{
create_sascfg_personal_py(local = TRUE, iomhost = 'aws-saspv1.mycompany.com', class_id = 'letter-number')
}   
}
