% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_pass.R
\name{delete_pass}
\alias{delete_pass}
\title{delete_pass}
\usage{
delete_pass(local = TRUE)
}
\arguments{
\item{local}{TRUE or FALSE either local windows machine or workbench remote}
}
\value{
"pass deleted"
}
\description{
delete the authinfo file to connect SAS
}
\examples{
\dontrun{
delete_pass(local = TRUE)
}   
}
