% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_python_module.R
\name{install_python_module}
\alias{install_python_module}
\title{install_python_module}
\usage{
install_python_module(moption = FALSE)
}
\arguments{
\item{moption}{TRUE or FALSE sometime it is required to specify python3 -m to install a module}
}
\value{
"Modules installed"
}
\description{
install python modules to connect SAS
}
\examples{
\dontrun{
install_python_module()
}    
}
