% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{SAS}
\alias{SAS}
\title{SAS}
\description{
SAS

SAS
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NA-new}{\code{SAS$new()}}
\item \href{#method-NA-submit}{\code{SAS$submit()}}
\item \href{#method-NA-submitLST}{\code{SAS$submitLST()}}
\item \href{#method-NA-get}{\code{SAS$get()}}
\item \href{#method-NA-to}{\code{SAS$to()}}
\item \href{#method-unknown-clone}{\code{SAS$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-new"></a>}}
\if{latex}{\out{\hypertarget{method-NA-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SAS$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-submit"></a>}}
\if{latex}{\out{\hypertarget{method-NA-submit}{}}}
\subsection{Method \code{submit()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SAS$submit(code)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{code}}{code chunk}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-submitLST"></a>}}
\if{latex}{\out{\hypertarget{method-NA-submitLST}{}}}
\subsection{Method \code{submitLST()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SAS$submitLST(code)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{code}}{code chunk}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SAS$get(table, lib)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{table}}{name of table in SAS}

\item{\code{lib}}{name of SAS libname}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-to"></a>}}
\if{latex}{\out{\hypertarget{method-NA-to}{}}}
\subsection{Method \code{to()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SAS$to(df, table, libref)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{dataframe R}

\item{\code{table}}{name of table in SAS}

\item{\code{libref}}{name of SAS libname}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-unknown-clone"></a>}}
\if{latex}{\out{\hypertarget{method-unknown-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SAS$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
